refactor(test): restructure zebrad tests into module-based architecture
Split the 4,440-line acceptance.rs monolith into a single test binary
with three module tiers: unit::, integration::, and stateful::. This
replaces 23 nextest profiles with 4 semantic profiles and eliminates
the manually maintained exclusion denylist.

Test categorization is now structural: adding a new test requires
placing a function in the right module file with zero config changes.
The nextest default-filter excludes stateful:: tests by module path,
removing the fragile per-test-name denylist.

Key changes:
- Split acceptance.rs into unit/, integration/, stateful/ modules
- Absorb config.rs and end_of_support.rs into unit/ modules
- Replace 23 nextest profiles with 4 (default, ci, ci-stateful,
  check-no-git-dependencies)
- Remove env var gates (TEST_LARGE_CHECKPOINTS, TEST_SYNC_TO_CHECKPOINT,
  TEST_SYNC_PAST_CHECKPOINT) from test functions
- Update Docker entrypoint to support NEXTEST_FILTER env var
- Update CI workflows to use new profiles and filter expressions
- Preserve feature flags for external deps (lightwalletd, checkpoints)
- Add ADR documenting the architectural decision

Closes #10136
