+++ zebrad/Cargo.toml
@@ -16,6 +16,7 @@ categories = ["command-line-utilities", "cryptography::cryptocurrencies"]
 
 # Settings that impact compilation
 edition.workspace = true
+autotests = false
 
 # Zebra can have a higher MSRV than the other crates since it is a binary.
 rust-version = "1.91"
@@ -48,6 +49,10 @@ features = [
     "sentry",
     "indexer",
 ]
+
+[[test]]
+name = "zebrad-tests"
+path = "tests/main.rs"
 
 [features]
 # In release builds, don't compile debug logging code, to improve performance.
