+++ zebrad/tests/common/mod.rs
@@ -17,6 +17,7 @@ pub mod coinbase;
 pub mod config;
 pub mod failure_messages;
 pub mod get_block_template_rpcs;
+#[macro_use]
 pub mod launch;
 pub mod lightwalletd;
 pub mod regtest;
