+++ zebrad/tests/e2e/lightwalletd.rs
@@ -1,0 +1,23 @@
+use color_eyre::eyre::Result;
+
+use crate::{
+    common::test_type::TestType::FullSyncFromGenesis, stateful::sync::lwd_integration_test,
+};
+
+/// Make sure `lightwalletd` can fully sync from genesis using Zebra.
+///
+/// This test only runs when:
+/// - `TEST_LIGHTWALLETD` is set,
+/// - a persistent cached state is configured (e.g., via `ZEBRA_STATE__CACHE_DIR`), and
+/// - Zebra is compiled with `--features=lightwalletd-grpc-tests`.
+///
+///
+/// This test doesn't work on Windows, so it is always skipped on that platform.
+#[test]
+#[ignore]
+#[cfg(not(target_os = "windows"))]
+fn lwd_sync_full() -> Result<()> {
+    lwd_integration_test(FullSyncFromGenesis {
+        allow_lightwalletd_cached_state: false,
+    })
+}
