+++ zebrad/tests/e2e/mod.rs
@@ -1,0 +1,7 @@
+pub mod sync;
+
+#[cfg(feature = "zebra-checkpoints")]
+pub mod checkpoints;
+
+#[cfg(feature = "lightwalletd-grpc-tests")]
+pub mod lightwalletd;
