+++ zebrad/tests/integration/mod.rs
@@ -1,0 +1,6 @@
+pub mod database;
+pub mod mempool;
+pub mod network;
+pub mod regtest;
+pub mod rpc;
+pub mod sync;
