+++ zebrad/tests/stateful/mod.rs
@@ -1,0 +1,6 @@
+#[cfg(feature = "indexer")]
+pub mod indexer;
+#[cfg(feature = "lightwalletd-grpc-tests")]
+pub mod lightwalletd;
+pub mod rpc;
+pub mod sync;
