+++ zebrad/tests/unit/cli.rs
@@ -1,0 +1,248 @@
+use std::time::Duration;
+
+use color_eyre::eyre::Result;
+
+use zebra_chain::parameters::Network::*;
+use zebra_test::{args, prelude::*};
+
+use crate::common::{
+    check::is_zebrad_version,
+    config::{default_test_config, external_address_test_config, persistent_test_config, testdir},
+    launch::{ZebradTestDirExt, LAUNCH_DELAY},
+};
+
+#[test]
+fn generate_no_args() -> Result<()> {
+    let _init_guard = zebra_test::init();
+
+    let child = testdir()?
+        .with_config(&mut default_test_config(&Mainnet)?)?
+        .spawn_child(args!["generate"])?;
+
+    let output = child.wait_with_output()?;
+    let output = output.assert_success()?;
+
+    // First line
+    output.stdout_line_contains("# Default configuration for zebrad")?;
+
+    Ok(())
+}
+
+#[test]
+fn generate_args() -> Result<()> {
+    let _init_guard = zebra_test::init();
+
+    let testdir = testdir()?;
+    let testdir = &testdir;
+
+    // unexpected free argument `argument`
+    let child = testdir.spawn_child(args!["generate", "argument"])?;
+    let output = child.wait_with_output()?;
+    output.assert_failure()?;
+
+    // unrecognized option `-f`
+    let child = testdir.spawn_child(args!["generate", "-f"])?;
+    let output = child.wait_with_output()?;
+    output.assert_failure()?;
+
+    // missing argument to option `-o`
+    let child = testdir.spawn_child(args!["generate", "-o"])?;
+    let output = child.wait_with_output()?;
+    output.assert_failure()?;
+
+    // Add a config file name to tempdir path
+    let generated_config_path = testdir.path().join("zebrad.toml");
+
+    // Valid
+    let child =
+        testdir.spawn_child(args!["generate", "-o": generated_config_path.to_str().unwrap()])?;
+
+    let output = child.wait_with_output()?;
+    let output = output.assert_success()?;
+
+    assert_with_context!(
+        testdir.path().exists(),
+        &output,
+        "test temp directory not found"
+    );
+    assert_with_context!(
+        generated_config_path.exists(),
+        &output,
+        "generated config file not found"
+    );
+
+    Ok(())
+}
+
+#[test]
+fn help_no_args() -> Result<()> {
+    let _init_guard = zebra_test::init();
+
+    let testdir = testdir()?.with_config(&mut default_test_config(&Mainnet)?)?;
+
+    let child = testdir.spawn_child(args!["help"])?;
+    let output = child.wait_with_output()?;
+    let output = output.assert_success()?;
+
+    // The first line should have the version
+    output.any_output_line(
+        is_zebrad_version,
+        &output.output.stdout,
+        "stdout",
+        "are valid zebrad semantic versions",
+    )?;
+
+    // Make sure we are in help by looking for the usage string
+    output.stdout_line_contains("Usage:")?;
+
+    Ok(())
+}
+
+#[test]
+fn help_args() -> Result<()> {
+    let _init_guard = zebra_test::init();
+
+    let testdir = testdir()?;
+    let testdir = &testdir;
+
+    // The subcommand "argument" wasn't recognized.
+    let child = testdir.spawn_child(args!["help", "argument"])?;
+    let output = child.wait_with_output()?;
+    output.assert_failure()?;
+
+    // option `-f` does not accept an argument
+    let child = testdir.spawn_child(args!["help", "-f"])?;
+    let output = child.wait_with_output()?;
+    output.assert_failure()?;
+
+    Ok(())
+}
+
+#[test]
+fn start_no_args() -> Result<()> {
+    let _init_guard = zebra_test::init();
+
+    // start caches state, so run one of the start tests with persistent state
+    let testdir = testdir()?.with_config(&mut persistent_test_config(&Mainnet)?)?;
+
+    let mut child = testdir.spawn_child(args!["-v", "start"])?;
+
+    // Run the program and kill it after a few seconds
+    std::thread::sleep(LAUNCH_DELAY);
+    child.kill(false)?;
+
+    let output = child.wait_with_output()?;
+    let output = output.assert_failure()?;
+
+    output.stdout_line_contains("Starting zebrad")?;
+
+    // Make sure the command passed the legacy chain check
+    output.stdout_line_contains("starting legacy chain check")?;
+    output.stdout_line_contains("no legacy chain found")?;
+
+    // Make sure the command was killed
+    output.assert_was_killed()?;
+
+    Ok(())
+}
+
+#[test]
+fn start_args() -> Result<()> {
+    let _init_guard = zebra_test::init();
+
+    let testdir = testdir()?.with_config(&mut default_test_config(&Mainnet)?)?;
+    let testdir = &testdir;
+
+    let mut child = testdir.spawn_child(args!["start"])?;
+    // Run the program and kill it after a few seconds
+    std::thread::sleep(LAUNCH_DELAY);
+    child.kill(false)?;
+    let output = child.wait_with_output()?;
+
+    // Make sure the command was killed
+    output.assert_was_killed()?;
+
+    output.assert_failure()?;
+
+    // unrecognized option `-f`
+    let child = testdir.spawn_child(args!["start", "-f"])?;
+    let output = child.wait_with_output()?;
+    output.assert_failure()?;
+
+    Ok(())
+}
+
+#[test]
+fn version_no_args() -> Result<()> {
+    let _init_guard = zebra_test::init();
+
+    let testdir = testdir()?.with_config(&mut default_test_config(&Mainnet)?)?;
+
+    let child = testdir.spawn_child(args!["--version"])?;
+    let output = child.wait_with_output()?;
+    let output = output.assert_success()?;
+
+    // The output should only contain the version
+    output.output_check(
+        is_zebrad_version,
+        &output.output.stdout,
+        "stdout",
+        "a valid zebrad semantic version",
+    )?;
+
+    Ok(())
+}
+
+#[test]
+fn version_args() -> Result<()> {
+    let _init_guard = zebra_test::init();
+
+    let testdir = testdir()?.with_config(&mut default_test_config(&Mainnet)?)?;
+    let testdir = &testdir;
+
+    // unrecognized option `-f`
+    let child = testdir.spawn_child(args!["tip-height", "-f"])?;
+    let output = child.wait_with_output()?;
+    output.assert_failure()?;
+
+    // unrecognized option `-f` is ignored
+    let child = testdir.spawn_child(args!["--version", "-f"])?;
+    let output = child.wait_with_output()?;
+    let output = output.assert_success()?;
+
+    // The output should only contain the version
+    output.output_check(
+        is_zebrad_version,
+        &output.output.stdout,
+        "stdout",
+        "a valid zebrad semantic version",
+    )?;
+
+    Ok(())
+}
+
+#[test]
+fn external_address() -> Result<()> {
+    let _init_guard = zebra_test::init();
+    let testdir = testdir()?.with_config(&mut external_address_test_config(&Mainnet)?)?;
+    let mut child = testdir.spawn_child(args!["start"])?;
+
+    // Give enough time to start connecting to some peers.
+    std::thread::sleep(Duration::from_secs(10));
+
+    child.kill(false)?;
+
+    let output = child.wait_with_output()?;
+    let output = output.assert_failure()?;
+
+    // Zebra started
+    output.stdout_line_contains("Starting zebrad")?;
+
+    // Make sure we are using external address for Version messages.
+    output.stdout_line_contains("using external address for Version messages")?;
+
+    // Make sure the command was killed.
+    output.assert_was_killed()?;
+
+    Ok(())
+}
