+++ zebrad/tests/unit/end_of_support.rs
@@ -1,16 +1,61 @@
-//! Testing the end of support feature.
-
-#![allow(clippy::unwrap_in_result)]
-
-use std::time::Duration;
+use std::{fs, time::Duration};
 
 use color_eyre::eyre::Result;
 
-use zebra_chain::{block::Height, chain_tip::mock::MockChainTip, parameters::Network};
+use zebra_chain::{
+    block::Height,
+    chain_tip::mock::MockChainTip,
+    parameters::{Network, Network::*},
+};
+use zebra_test::{args, prelude::*};
 use zebrad::components::sync::end_of_support::{self, EOS_PANIC_AFTER, ESTIMATED_RELEASE_HEIGHT};
+
+use crate::common::{
+    config::{default_test_config, testdir},
+    launch::ZebradTestDirExt,
+};
 
 // Estimated blocks per day with the current 75 seconds block spacing.
 const ESTIMATED_BLOCKS_PER_DAY: u32 = 1152;
+
+/// Check that the end of support code is called at least once.
+#[test]
+fn end_of_support_is_checked_at_start() -> Result<()> {
+    let _init_guard = zebra_test::init();
+    let testdir = testdir()?.with_config(&mut default_test_config(&Mainnet)?)?;
+    let mut child = testdir.spawn_child(args!["start"])?;
+
+    // Give enough time to start up the eos task.
+    std::thread::sleep(Duration::from_secs(30));
+
+    child.kill(false)?;
+
+    let output = child.wait_with_output()?;
+    let output = output.assert_failure()?;
+
+    // Zebra started
+    output.stdout_line_contains("Starting zebrad")?;
+
+    // End of support task started.
+    output.stdout_line_contains("Starting end of support task")?;
+
+    // Make sure the command was killed
+    output.assert_was_killed()?;
+
+    Ok(())
+}
+
+/// Check that Zebra does not depend on any crates from git sources.
+#[test]
+#[ignore]
+fn check_no_git_dependencies() {
+    let cargo_lock_contents =
+        fs::read_to_string("../Cargo.lock").expect("should have Cargo.lock file in root dir");
+
+    if cargo_lock_contents.contains(r#"source = "git+"#) {
+        panic!("Cargo.lock includes git sources")
+    }
+}
 
 /// Test that the `end_of_support` function is working as expected.
 #[test]
